Raylan - DDS 1-12

+++ src/exerciciosPythonBrasil_EstruturaDeDecisao/Exercicio01.java
@@ -6,11 +6,10 @@ public class Exercicio01 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Digite um número: ");
-        double primeiroNumero = scanner.nextDouble();
+        float primeiroNumero = scanner.nextFloat();
         System.out.print("Digite outro número: ");
-        double segundoNumero = scanner.nextDouble();
+        float segundoNumero = scanner.nextFloat();
         boolean maiorNumero = primeiroNumero > segundoNumero;
-
         if (maiorNumero) {
             System.out.println("O maior número é: " + primeiroNumero);
         } else {
